backend/requirements.txt
amqp==5.0.6
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.26.0
colorama==0.4.4
coverage==5.5
crispy-tailwind==0.5.0
cryptography==35.0.0
dj-database-url==0.5.0
django-autoslug==1.9.8
django-celery-results==2.2.0
django-countries==7.2.1
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-heroku==0.3.1
django-recaptcha==2.0.6
django-taggit==1.5.1
djangorestframework==3.15.1
flake8==3.9.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
isort==5.9.3
kombu==5.1.0
mccabe==0.6.1
outcome==1.1.0
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.8.0
pyOpenSSL==21.0.0
python-dotenv==0.19.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
selenium==4.0.0
sentry-sdk==1.4.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.1
stripe==2.61.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.6
uvicorn==0.15.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wsproto==1.0.0
